Setupactivity-n alaplap rész hozzáadása, socket alapján szűrés cpu-ra.

+++ app/src/main/java/com/example/hf/models/CpuPart.java
@@ -3,13 +3,13 @@ package com.example.hf.models;
 public class CpuPart {
     private String name;
     private String socket;
-    private long core;
+    private int core;
 
     public CpuPart() {
         // Üres konstruktor szükséges Firestore-hoz
     }
 
-    public CpuPart(String name, String socket, long core) {
+    public CpuPart(String name, String socket, int core) {
         this.name = name;
         this.socket = socket;
         this.core = core;
@@ -18,12 +18,10 @@ public class CpuPart {
     public String getName() {
         return name;
     }
-
     public String getSocket() {
         return socket;
     }
-
-    public long getCore() {
+    public int getCore() {
         return core;
     }
 }
